// pop-web/src/app/shared/openapi/model/radiotherapy.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { User } from './user';
import { CodedConcept } from './coded-concept';
import { RadiotherapyDosage } from './radiotherapy-dosage';
import { Period } from './period';
import { RadiotherapyIntentChoices } from './radiotherapy-intent-choices';
import { RadiotherapySetting } from './radiotherapy-setting';


export interface Radiotherapy { 
    /**
     * Radiation doses administered during the radiotherapy
     */
    dosages: Array<RadiotherapyDosage>;
    /**
     * Settings of the radiotherapy irradiation procedure
     */
    settings: Array<RadiotherapySetting>;
    /**
     * 
     */
    id: string;
    /**
     * 
     */
    createdAt: string;
    /**
     * 
     */
    updatedAt: string;
    createdBy?: User | null;
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who received the radiotherapy
     */
    caseId: string;
    /**
     * Clinically-relevant period during which the radiotherapy was administered to the patient.
     */
    period: Period;
    /**
     * The total number of radiotherapy sessions over the treatment period.
     */
    sessions: number;
    /**
     * Treatment intent of the system therapy
     */
    intent: RadiotherapyIntentChoices;
    terminationReason?: CodedConcept | null;
    therapyLineId?: string | null;
    /**
     * The user(s) who updated the data since its creation
     */
    updatedBy?: Array<User>;
    /**
     * References to the neoplastic entities that were targeted by the radiotherapy
     */
    targetedEntitiesIds?: Array<string>;
    /**
     * Human-readable description
     */
    description: string;
}
export namespace Radiotherapy {
}